performance optimization

+++ source/EngineGpuKernels/TOs.cuh
@@ -12,17 +12,17 @@ struct ParticleTO
 	float energy;
 	float2 pos;
 	float2 vel;
-    int color;
+    uint8_t color;
 
-	int selected;
+	uint8_t selected;
 };
 
 struct CellMetadataTO
 {
-    int nameSize;
+    uint16_t nameSize;
     uint64_t nameDataIndex;
 
-    int descriptionSize;
+    uint16_t descriptionSize;
     uint64_t descriptionDataIndex;
 };
 
@@ -51,21 +51,21 @@ struct TransmitterTO
 
 struct ConstructorTO
 {
-    int activationMode;  //0 = manual, 1 = every cycle, 2 = every second cycle, 3 = every third cycle, etc.
-    int constructionActivationTime;
+    uint32_t activationMode;  //0 = manual, 1 = every cycle, 2 = every second cycle, 3 = every third cycle, etc.
+    uint32_t constructionActivationTime;
 
-    int genomeSize;
+    uint16_t genomeSize;
     uint64_t genomeDataIndex;
-    int genomeGeneration;
+    uint32_t genomeGeneration;
     float constructionAngle1;
     float constructionAngle2;
 
     //process data
     uint64_t lastConstructedCellId;
-    int genomeCurrentNodeIndex;
-    int genomeCurrentRepetition;
-    int offspringCreatureId;
-    int offspringMutationId;
+    uint16_t genomeCurrentNodeIndex;
+    uint16_t genomeCurrentRepetition;
+    uint32_t offspringCreatureId;
+    uint32_t offspringMutationId;
     uint32_t stateFlags;  //bit 0: isConstructionBuilt, bit 1: isInherited
 };
 
@@ -74,8 +74,8 @@ struct SensorTO
     SensorMode mode;
     float angle;
     float minDensity;
-    int color;
-    int targetedCreatureId;
+    uint8_t color;
+    uint32_t targetedCreatureId;
 
     //process data
     float memoryChannel1;
@@ -85,8 +85,8 @@ struct SensorTO
 
 struct NerveTO
 {
-    int pulseMode;        //0 = none, 1 = every cycle, 2 = every second cycle, 3 = every third cycle, etc.
-    int alternationMode;  //0 = none, 1 = alternate after each pulse, 2 = alternate after second pulse, 3 = alternate after third pulse, etc.
+    uint8_t pulseMode;    //0 = none, 1 = every cycle, 2 = every second cycle, 3 = every third cycle, etc.
+    uint8_t alternationMode;  //0 = none, 1 = alternate after each pulse, 2 = alternate after second pulse, 3 = alternate after third pulse, etc.
 };
 
 struct AttackerTO
@@ -97,17 +97,17 @@ struct AttackerTO
 struct InjectorTO
 {
     InjectorMode mode;
-    int counter;
-    int genomeSize;
+    uint32_t counter;
+    uint16_t genomeSize;
     uint64_t genomeDataIndex;
-    int genomeGeneration;
+    uint32_t genomeGeneration;
 };
 
 struct MuscleTO
 {
     MuscleMode mode;
     MuscleBendingDirection lastBendingDirection;
-    int lastBendingSourceIndex;
+    uint8_t lastBendingSourceIndex;
     float consecutiveBendingAngle;
 };
 
@@ -118,13 +118,13 @@ struct DefenderTO
 
 struct ReconnectorTO
 {
-    int color;
+    uint8_t color;
 };
 
 struct DetonatorTO
 {
     DetonatorState state;
-    int countdown;
+    int32_t countdown;
 };
 
 union CellFunctionTO
@@ -152,28 +152,28 @@ struct CellTO
     float2 vel;
 	float energy;
     float stiffness;
-    int color;
-    int maxConnections;
-	int numConnections;
+    uint8_t color;
+    uint8_t maxConnections;
+    uint8_t numConnections;
     bool barrier;
-    int age;
+    uint32_t age;
     LivingState livingState;
-    int creatureId;
-    int mutationId;
+    uint32_t creatureId;
+    uint32_t mutationId;
 
     //cell function
-    int executionOrderNumber;
-    int inputExecutionOrderNumber;
+    uint8_t executionOrderNumber;
+    int8_t inputExecutionOrderNumber;
     bool outputBlocked;
     CellFunction cellFunction;
     CellFunctionTO cellFunctionData;
     ActivityTO activity;
-    int activationTime;
-    int genomeNumNodes;
+    uint32_t activationTime;
+    uint32_t genomeNumNodes;
 
     CellMetadataTO metadata;
 
-    int selected;
+    uint8_t selected;
 };
 
 struct DataTO
